refactor: rename api-url file to api file
